@@ -14,37 +14,70 @@
   },
   "dependencies": {
     "@clerk/nextjs": "^6.34.0",
+    "@heroui/avatar": "^2.2.22",
+    "@heroui/badge": "^2.2.17",
     "@heroui/button": "^2.2.27",
     "@heroui/card": "^2.2.25",
+    "@heroui/code": "^2.2.21",
     "@heroui/divider": "^2.2.20",
+    "@heroui/dropdown": "^2.3.27",
     "@heroui/input": "^2.4.28",
+    "@heroui/kbd": "^2.2.22",
+    "@heroui/link": "^2.2.23",
+    "@heroui/listbox": "^2.3.26",
+    "@heroui/modal": "^2.2.24",
+    "@heroui/navbar": "^2.2.25",
+    "@heroui/progress": "^2.2.22",
+    "@heroui/snippet": "^2.2.28",
+    "@heroui/spinner": "^2.2.24",
+    "@heroui/switch": "^2.2.24",
     "@heroui/system": "^2.4.23",
+    "@heroui/table": "^2.2.27",
+    "@heroui/tabs": "^2.2.24",
     "@heroui/theme": "^2.4.23",
+    "@heroui/toast": "^2.0.17",
+    "@heroui/tooltip": "^2.2.24",
     "@hookform/resolvers": "^5.2.2",
     "@imagekit/next": "^2.1.3",
     "@neondatabase/serverless": "^1.0.2",
+    "axios": "^1.13.1",
+    "clsx": "^2.1.1",
+    "date-fns": "^4.1.0",
     "dotenv": "^17.2.3",
     "drizzle-orm": "^0.44.7",
     "framer-motion": "^12.23.24",
     "imagekit": "^6.0.0",
+    "imagekitio-next": "^1.0.1",
+    "intl-messageformat": "^10.7.18",
     "lucide-react": "^0.548.0",
     "next": "16.0.0",
+    "next-themes": "^0.4.6",
     "react": "19.2.0",
     "react-dom": "19.2.0",
     "react-hook-form": "^7.65.0",
-    "tsx": "^4.20.6",
     "uuid": "^13.0.0",
     "zod": "^3.25.76"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
-    "drizzle-kit": "^0.31.5",
-    "eslint": "^9",
-    "eslint-config-next": "16.0.0",
-    "tailwindcss": "^4",
-    "typescript": "^5"
+    "@tailwindcss/postcss": "^4.1.16",
+    "@types/node": "^20.19.24",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2",
+    "@typescript-eslint/eslint-plugin": "^8.46.2",
+    "@typescript-eslint/parser": "^8.46.2",
+    "autoprefixer": "^10.4.21",
+    "drizzle-kit": "^0.31.6",
+    "eslint": "^9.38.0",
+    "eslint-config-next": "^16.0.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "eslint-plugin-unused-imports": "^4.3.0",
+    "postcss": "^8.5.6",
+    "prettier": "^3.6.2",
+    "tailwind-variants": "^3.1.1",
+    "tailwindcss": "^4.1.16",
+    "tsx": "^4.20.6",
+    "typescript": "^5.9.3"
   }
 }
