@@ -1,26 +1,28 @@
-import { clerkMiddleware, createRouteMatcher, auth } from "@clerk/nextjs/server";
-import { NextResponse } from "next/server";
-
-
-
-const isPublicRoute = createRouteMatcher(["/", "sign-in(.*)","sign-up(.*)"])
-
-export default clerkMiddleware(async (auth, request) => {
-    const user = auth()
-    const userId = (await user).userId
-    const url = new URL(request.url)
-
-    if(userId && isPublicRoute(request) && url.pathname !== "/" ){
-        return NextResponse.redirect(new URL("/dashboard", request.url))
+import {
+    clerkMiddleware,
+    createRouteMatcher,
+    auth,
+  } from "@clerk/nextjs/server";
+  import { NextResponse } from "next/server";
+  
+  const isPublicRoute = createRouteMatcher(["/", "/sign-in(.*)", "/sign-up(.*)"]);
+  
+  export default clerkMiddleware(async (auth, request) => {
+    const user = auth();
+    const userId = (await user).userId;
+    const url = new URL(request.url);
+  
+    if (userId && isPublicRoute(request) && url.pathname !== "/") {
+      return NextResponse.redirect(new URL("/dashboard", request.url));
     }
-
-    //protect non-public routes
-    if(!isPublicRoute(request)){
-        await auth.protect()
+  
+    // Protect non-public routes
+    if (!isPublicRoute(request)) {
+      await auth.protect();
     }
-});
-
-export const config = {
+  });
+  
+  export const config = {
     matcher: [
       // Skip Next.js internals and all static files, unless found in search params
       "/((?!_next|[^?]*\\.(?:html?|css|js(?!on)|jpe?g|webp|png|gif|svg|ttf|woff2?|ico|csv|docx?|xlsx?|zip|webmanifest)).*)",
