@@ -1,7 +1,6 @@
-const config = {
+export default {
   plugins: {
     "@tailwindcss/postcss": {},
+    autoprefixer: {},
   },
 };
-
-export default config;
